requirements.txt
#create a virtual environment with these packages in order to run this project
ipython==7.5.0
pytest==4.5.0
pytest-html==1.20.0
pytest-ordering==0.6
selenium==3.141.0
ddt==1.2.1
openpyxl=2.6.2
pytest-xdist=1.28.0
Appium-Python-Client=0.44
path.py=12.0.1
